JUNIPER-ONLY: Read-only User should not be able to delete ICMP/SNMP service of one node for 935273

+++ opennms-webapp/src/main/webapp/WEB-INF/jsp/element/service.jsp
@@ -148,11 +148,11 @@ function doDelete() {
       <p>
          <a href="${eventUrl}">View Events</a>
          
- 	
+ <% if( request.isUserInRole( Authentication.ROLE_ADMIN ) || !request.isUserInRole( Authentication.ROLE_READONLY ) ) { %>	
        <sec:authorize url="admin/deleteService">
          &nbsp;&nbsp;&nbsp;<a href="admin/deleteService" onClick="return doDelete()">Delete</a>
        </sec:authorize>
-
+<% } %>
 	
 
       </p>
